Fix front-end tests
Content in front-end changes multiple times, as patches are sent
to the backend. It's not reliable to wait for the first change,
which is sometimes an undefined error (@dralletje??!!)
so we wait until the result stabilizes to the value we expect.

To the user that's transparent (hopefully). We've also added
a counter to keep track of pending backend requests, and add a
class to the document.body (that doesn't do anything else).