blabla (#881)
* Scope Firebasey stuffy

* Run workflows in this branch

* Add 'Nothing' as possible expression

This happens when you interrupt a cell.

* Adding immer to notebook diff

That should not be needed. The assumption is that
the state becomes immutable/frozen after an immer call,
so mutations don't play well with the returned object.
immer should return a full object so that should not
emerge. Testing will tell.

* Disable immer autofreeze

* Comment setAutoFreeze so a ninja can fix properly in the future

* update immer comment

* Revert "Add 'Nothing' as possible expression"

This reverts commit 21d0ecdb676a2d3ac5c2936559b2a636f448c3a8.

* Fix drag & drop cells to other page

The issue here was that the data doesn't flow
from cell input to the localcode.

Update the codemirror using DOM. Another (ugly)
solution would be to add a useEffect but since
this is not handled by React anyway, I don't suggest
it.

Co-authored-by: Panagiotis Georgakopoulos <panagiotis.georgakopoulos@intelligencia.ai>