prevent scrollbar on dataframe output (#1261)
Any dataframe with more than 12 rows automatically has a scrollbar in Pluto since it is more than 400px high. 450px is enough to fit dataframes with more than 15 rows since dataframes automatically hides rows. 502px is enough to show up to 15 rows without scrollbar.

I mentioned the scrollbar before [here](https://github.com/fonsp/Pluto.jl/issues/996), but was told is was a feature so this PR might not be accepted. But I thought I give it a try anyway.