Simplify tuple handling in ExpressionExplorer
The Julia syntax `:(a, b = 1, 2)` is not a multiple assignment to a and b
but the constructor for a named tuple (which can be confusing) with fields a, b and 2 (which is
invalid). The conditional handling for multiple assignments and named
tuple can then be removed because multiple assignments will always be
top-level `:(=)`, instead of `:tuple` as though previously.