Add stale cells (aka single-shot cells)
Adds to PlutoRunner a new struct `Stale`. When returned from a cell,
this struct passes down its formatting to its single member `out`.
However, the cell also gets marked as stale. This means that none of the
cells that depend on it will be run. The stale cell, however, will still
be run when its dependencies run.

On the frontend stale outputs are displayed on a light or dark orange
background while stale cells and cells that depend on them get orange
traffic lights.

When a cell is in one of those two states a warning sign with an
explanatory popup is displayed below the "fold code" button.

Add a small test for stale cells