Write a short docstring...
that will be printed in the beginning of the help string.
Notice that the formatting (i.e., the line breaks) are sometimes a bit
scuffed, but that seems to be a limitation of Markdown.jl as discussed
in https://github.com/comonicon/Comonicon.jl/issues/213#issuecomment-1212520305.

Notice also that the `# Intro` header is required for Comonicon to
accept this docstring (otherwise it is skipped) as per https://comonicon.org/stable/conventions/#Description.